refactor(user): move update-user.dto to user/dtos directory and update imports
This change reorganizes the user-related DTOs by moving UpdateUserDto to a more appropriate location under domain/user/dtos. All import paths referencing the old location have been updated accordingly.
