+++ src/config/__tests__/mocks/auth-repository.mock.ts
@@ -2,7 +2,7 @@ import { LoginDto } from "@/domain/auth/dtos/login.dto";
 import { LogoutDto } from "@/domain/auth/dtos/logout.dto";
 import { RegisterDto } from "@/domain/auth/dtos/register.dto";
 import { RequestResetPasswordEmailDto } from "@/domain/auth/dtos/request-reset-password-email.dto";
-import { UpdateUserDto } from "@/domain/dtos/update-user.dto";
+import { UpdateUserDto } from "@/domain/user/dtos/update-user.dto";
 import { AuthUserEntity } from "@/domain/entities/auth-user.entity";
 import { UserEntity } from "@/domain/entities/user.entity";
 import { AuthRepository } from "@/domain/auth/repositories/auth.repository";
