+++ src/domain/auth/repositories/__test__/auth.repository.test.ts
@@ -5,7 +5,7 @@ import { UserEntity } from "../../../entities/user.entity";
 import { LoginDto } from "@/domain/auth/dtos/login.dto";
 import { LogoutDto } from "@/domain/auth/dtos/logout.dto";
 import { RequestResetPasswordEmailDto } from "@/domain/auth/dtos/request-reset-password-email.dto";
-import { UpdateUserDto } from "@/domain/dtos/update-user.dto";
+import { UpdateUserDto } from "@/domain/user/dtos/update-user.dto";
 
 // Concrete implementation for testing
 class TestAuthRepository extends AuthRepository {
