+++ src/domain/use-cases/__tests__/update-user-password.test.ts
@@ -1,5 +1,5 @@
 import { UpdateUserPassword } from "../update-user-password";
-import { UpdateUserDto } from "../../dtos/update-user.dto";
+import { UpdateUserDto } from "../../user/dtos/update-user.dto";
 import { UserEntity } from "../../entities/user.entity";
 import { AuthUserEntity } from "../../entities/auth-user.entity";
 import { MockAuthRepository } from "@/config/__tests__/mocks/auth-repository.mock";
